requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6